Inline form hooks
