+++ frontend/src/metabase/core/components/FormSubmitButton/FormSubmitButton.tsx
@@ -1,9 +1,14 @@
-import React, { forwardRef, Ref } from "react";
+import React, {
+  forwardRef,
+  Ref,
+  useEffect,
+  useLayoutEffect,
+  useState,
+} from "react";
 import { useFormikContext } from "formik";
 import { t } from "ttag";
 import Button, { ButtonProps } from "metabase/core/components/Button";
-import { FormStatus } from "metabase/core/hooks/use-form-state";
-import useFormStatus from "metabase/core/hooks/use-form-status";
+import useFormState from "metabase/core/hooks/use-form-state";
 
 export interface FormSubmitButtonProps extends Omit<ButtonProps, "children"> {
   normalText?: string;
@@ -36,8 +41,38 @@ const FormSubmitButton = forwardRef(function FormSubmitButton(
   );
 });
 
+const STATUS_TIMEOUT = 5000;
+
+const useFormStatus = () => {
+  const { status } = useFormState();
+  const isRecent = useIsRecent(status, STATUS_TIMEOUT);
+
+  switch (status) {
+    case "pending":
+      return status;
+    case "fulfilled":
+    case "rejected":
+      return isRecent ? status : undefined;
+  }
+};
+
+const useIsRecent = (value: unknown, timeout: number) => {
+  const [isRecent, setIsRecent] = useState(true);
+
+  useEffect(() => {
+    const timerId = setTimeout(() => setIsRecent(false), timeout);
+    return () => clearTimeout(timerId);
+  }, [value, timeout]);
+
+  useLayoutEffect(() => {
+    setIsRecent(true);
+  }, [value]);
+
+  return isRecent;
+};
+
 const getSubmitButtonText = (
-  status: FormStatus | undefined,
+  status: "pending" | "fulfilled" | "rejected" | undefined,
   {
     normalText = t`Submit`,
     activeText = normalText,
