--- frontend/src/metabase/core/hooks/use-form-error-message/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./use-form-error-message";
