--- frontend/src/metabase/core/hooks/use-form-status/index.ts
@@ -1,1 +1,0 @@
-export { default } from "./use-form-status";
